requirements-dev.txt
django-functest~=1.5.4
pre-commit~=3.5.0
